onepiece-tcg-monitor/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pyyaml==6.0.1
apscheduler==3.10.4
python-dateutil==2.8.2
